Cleaned up schemaValidator and fixed bug in autocompleter
